Migrate to BigInts
Migrate the database from Int/FK to BigInt/FK and refactor relevant code to support this change.

Approved-by: Shibli Saleheen <ssaleheen@swin.edu.au>

Merged by: Lewis Lakerink <llakerink@swin.edu.au>

See merge request https://gitlab.com/CAS-eResearch/GWDC/code/gwcloud_job_server/-/merge_requests/34

+++ src/Cluster/Cluster.cpp
@@ -321,7 +321,7 @@ void Cluster::updateJob(Message &message) {
                 select(all_of(jobTable))
                         .from(jobTable)
                         .where(
-                                jobTable.id == static_cast<uint32_t>(jobId)
+                                jobTable.id == static_cast<uint64_t>(jobId)
                         )
         );
