+++ src/HTTP/tests/file_tests.cpp
@@ -10,9 +10,9 @@
 
 struct FileTestDataFixture : public DatabaseFixture, public HttpServerFixture, public HttpClientFixture {
     // NOLINTBEGIN(misc-non-private-member-variables-in-classes)
-    uint32_t jobId1;
-    uint32_t jobId2;
-    uint32_t jobId3;
+    uint64_t jobId1;
+    uint64_t jobId2;
+    uint64_t jobId3;
     // NOLINTEND(misc-non-private-member-variables-in-classes)
 
     FileTestDataFixture() {
